django-api/requirements.txt
Django==4.2.5
psycopg2_binary==2.9.7
djangorestframework===3.14.0
django-filter==23.2